新規投稿ページ_修正
