+++ src/app/_components/Wrapper/index.tsx
@@ -8,7 +8,7 @@ interface WrapperProps {
   children: React.ReactNode;
 }
 
-const Wrapper: React.FC<WrapperProps> = ({ size = 1000, children }) => {
+export const Wrapper: React.FC<WrapperProps> = ({ size = 1000, children }) => {
 
   const Size = () => {
     switch (size) {
@@ -19,19 +19,17 @@ const Wrapper: React.FC<WrapperProps> = ({ size = 1000, children }) => {
       case 800:
         return styles.wrapper800;
       case 900:
-        return  styles.wrapper900;
+        return styles.wrapper900;
       case 1000:
         return styles.wrapper1000;
       default:
         return styles.wrapper1000;
-    }
-  }
+    };
+  };
 
   return (
     <div className={Size()}>
       {children}
     </div>
-  )
-}
-
-export default Wrapper;
+  );
+};
