+++ src/app/api/tags/route.ts
@@ -72,8 +72,9 @@ export const POST = async (request: NextRequest) => {
     return NextResponse.json({
       status: 'OK',
       message: '作成しました',
-      id: data.id,
+      name: data.name,
     })
+
   } catch (error) {
     if (error instanceof Error) {
       return NextResponse.json({ status: error.message }, { status: 400 })
