+++ src/app/mypage/_styles/Mypage.module.scss
@@ -27,7 +27,7 @@
       font-size: 180%;
       font-weight: 900;
       padding-top: clamp(4rem, 4vw, 6rem);
-      padding-bottom: clamp(6rem, 8vw, 10rem);
+      padding-bottom: clamp(4rem, 4vw, 6rem);
     }
   }
 }
