+++ src/app/mypage/posts/new/_components/CategoryList/index.module.scss
@@ -1,0 +1,27 @@
+.selectArea {
+  margin-bottom: 3rem;
+  .category {
+    display: flex;
+    justify-content: start;
+    ul {
+      display: flex;
+      justify-content: start;
+      li {
+        margin-right: 1rem;
+        button {
+          background-color: #fff;
+          border: 1px solid #111;
+          border-radius: 3px;
+          color: #111;
+          padding: 0.2em 0.5em 0.3em;
+          &.selected {
+            background-color: #111;
+            border: 1px solid #111;
+            border-radius: 3px;
+            color: #fff;
+          }
+        }
+      }
+    }
+  }
+}
