+++ src/app/mypage/posts/new/_components/ReturnTop/index.module.scss
@@ -1,0 +1,9 @@
+.topLink {
+  margin-bottom: 2rem;
+  a {
+    border-bottom: 1px solid #111;
+    color: #111;
+    text-decoration: none;
+    padding-bottom: 0.1em;
+  }
+}
