+++ src/app/mypage/posts/new/_types/CustomElement.ts
@@ -1,0 +1,7 @@
+import { CustomText } from './CustomText';
+
+// カスタムエレメント型の定義
+export type CustomElement = {
+  type: 'paragraph';
+  children: CustomText[];
+}
